+++ activerecord/test/cases/tasks/database_tasks_test.rb
@@ -340,20 +340,6 @@ module ActiveRecord
       ENV["SCHEMA_CACHE"] = old_path
     end
 
-    def test_deprecated_cache_dump_default_filename
-      old_path = ENV["SCHEMA_CACHE"]
-      ENV.delete("SCHEMA_CACHE")
-
-      ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "db") do
-        path = assert_deprecated(ActiveRecord.deprecator) do
-          ActiveRecord::Tasks::DatabaseTasks.cache_dump_filename("primary")
-        end
-        assert_equal "db/schema_cache.yml", path
-      end
-    ensure
-      ENV["SCHEMA_CACHE"] = old_path
-    end
-
     def test_cache_dump_alternate_filename
       old_path = ENV["SCHEMA_CACHE"]
       ENV.delete("SCHEMA_CACHE")
@@ -362,20 +348,6 @@ module ActiveRecord
 
       ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "db") do
         path = ActiveRecord::Tasks::DatabaseTasks.cache_dump_filename(config)
-        assert_equal "db/alternate_schema_cache.yml", path
-      end
-    ensure
-      ENV["SCHEMA_CACHE"] = old_path
-    end
-
-    def test_deprecated_cache_dump_alternate_filename
-      old_path = ENV["SCHEMA_CACHE"]
-      ENV.delete("SCHEMA_CACHE")
-
-      ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "db") do
-        path = assert_deprecated(ActiveRecord.deprecator) do
-          ActiveRecord::Tasks::DatabaseTasks.cache_dump_filename("alternate")
-        end
         assert_equal "db/alternate_schema_cache.yml", path
       end
     ensure
@@ -391,20 +363,6 @@ module ActiveRecord
       ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "db") do
         path = assert_deprecated(/Setting `ENV\["SCHEMA_CACHE"\]` is deprecated and will be removed in Rails 8\.0\. Configure the `:schema_cache_path` in the database configuration instead\. \(/, ActiveRecord.deprecator) do
           ActiveRecord::Tasks::DatabaseTasks.cache_dump_filename(config)
-        end
-        assert_equal "db/schema_cache.yml", path
-      end
-    ensure
-      ENV["SCHEMA_CACHE"] = old_path
-    end
-
-    def test_deprecated_cache_dump_filename_with_env_override
-      old_path = ENV["SCHEMA_CACHE"]
-      ENV["SCHEMA_CACHE"] = "tmp/something.yml"
-
-      ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "db") do
-        path = assert_deprecated(/Passing a database name to `cache_dump_filename` is deprecated and will be removed in Rails 8\.0\. Pass a `ActiveRecord::DatabaseConfigurations::DatabaseConfig` object instead\. \(/, ActiveRecord.deprecator) do
-          ActiveRecord::Tasks::DatabaseTasks.cache_dump_filename("primary")
         end
         assert_equal "db/schema_cache.yml", path
       end
@@ -436,20 +394,6 @@ module ActiveRecord
       ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "db") do
         path = ActiveRecord::Tasks::DatabaseTasks.cache_dump_filename(config, schema_cache_path: "tmp/another.yml")
         assert_equal "tmp/another.yml", path
-      end
-    ensure
-      ENV["SCHEMA_CACHE"] = old_path
-    end
-
-    def test_deprecated_cache_dump_filename_with_path_from_the_argument
-      old_path = ENV["SCHEMA_CACHE"]
-      ENV.delete("SCHEMA_CACHE")
-
-      ActiveRecord::Tasks::DatabaseTasks.stub(:db_dir, "db") do
-        path = assert_deprecated(ActiveRecord.deprecator) do
-          ActiveRecord::Tasks::DatabaseTasks.cache_dump_filename("primary", schema_cache_path: "tmp/something.yml")
-        end
-        assert_equal "tmp/something.yml", path
       end
     ensure
       ENV["SCHEMA_CACHE"] = old_path